+++ src/widgets/link.rs
@@ -195,7 +195,7 @@ impl Widget for Link {
                 style = style.with_color(color.to_simple_opaque());
             }
             if let Some(bg) = resolved.link_background_hover.or(resolved.link_background) {
-                if bg.a > 0 {
+                if bg.a > 0.0 {
                     style = style.with_bgcolor(bg.to_simple_opaque());
                 }
             }
@@ -208,7 +208,7 @@ impl Widget for Link {
                 style = style.with_color(color.to_simple_opaque());
             }
             if let Some(bg) = resolved.link_background {
-                if bg.a > 0 {
+                if bg.a > 0.0 {
                     style = style.with_bgcolor(bg.to_simple_opaque());
                 }
             }